Removing parse.py and parse_settings
